+++ pyembed/core/consumer.py
@@ -20,7 +20,8 @@
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 # THE SOFTWARE.
 
-from pyembed.core import discovery, parse, render, PyEmbedError
+from pyembed.core import discovery, parse
+from pyembed.core.error import PyEmbedError
 
 import requests
 
@@ -28,28 +29,6 @@ import requests
 class PyEmbedConsumerError(PyEmbedError):
 
     """Thrown if there is an error discovering an OEmbed URL."""
-
-
-def embed(url,
-          max_width=None,
-          max_height=None,
-          renderer=None):
-    """Returns an HTML representation of a resource, given a URL.  This can be
-       directly embedded in a web page.
-
-    :param url: the content URL.
-    :param max_width: (optional) the maximum width of the embedded resource.
-    :param max_height: (optional) the maximum height of the embedded resource.
-    :param renderer: (optional) renderer to render the response.
-    :returns: an HTML representation of the resource.
-    :raises PyEmbedError: if there is an error fetching the response.
-    """
-    response = get_oembed_response(url, max_width, max_height)
-
-    if renderer is None:
-        renderer = render.DefaultRenderer()
-
-    return renderer.render(url, response)
 
 
 def get_oembed_response(url, max_width=None, max_height=None):