+++ pyembed/core/parse.py
@@ -20,8 +20,8 @@
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 # THE SOFTWARE.
 
-from pyembed.core import PyEmbedError
-from . import response
+from pyembed.core.error import PyEmbedError
+from pyembed.core import response
 
 from bs4 import BeautifulSoup
 import json