+++ pyembed/core/test/integration_test.py
@@ -20,13 +20,14 @@
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 # THE SOFTWARE.
 
-from pyembed.core import consumer, render
+from pyembed.core import PyEmbed
+from pyembed.core.render import PyEmbedRenderer
 
 from hamcrest import assert_that, contains_string, equal_to
 import pytest
 
 
-class DummyRenderer(render.PyEmbedRenderer):
+class DummyRenderer(PyEmbedRenderer):
 
     def render(self, content_url, response):
         return "%s by %s from %s" % \
@@ -34,27 +35,26 @@ class DummyRenderer(render.PyEmbedRenderer):
 
 
 def test_should_get_correct_embedding():
-    embedding = consumer.embed(
+    embedding = PyEmbed().embed(
         'https://twitter.com/BarackObama/status/266031293945503744')
     assert_that(embedding, contains_string('Four more years.'))
 
 
 def test_should_get_another_correct_embedding():
-    embedding = consumer.embed(
+    embedding = PyEmbed().embed(
         'http://www.flickr.com/photos/hansjuul/7899334594')
     assert_that(embedding, contains_string('.jpg'))
 
 
 def test_should_embed_with_maximum_height():
-    embedding = consumer.embed(
+    embedding = PyEmbed().embed(
         'http://www.youtube.com/watch?v=9bZkp7q19f0', max_height=200)
     assert_that(embedding, contains_string('height="200"'))
 
 
 def test_should_embed_with_custom_renderer():
-    embedding = consumer.embed(
-        'http://www.youtube.com/watch?v=qrO4YZeyl0I',
-        renderer=DummyRenderer())
+    embedding = PyEmbed(renderer=DummyRenderer()).embed(
+        'http://www.youtube.com/watch?v=qrO4YZeyl0I')
     assert_that(embedding, equal_to(
         'Lady Gaga - Bad Romance by LadyGagaVEVO from ' +
         'http://www.youtube.com/watch?v=qrO4YZeyl0I'))